finalização do app

+++ app/src/main/java/com/example/uana/model/Usuario.kt
@@ -4,14 +4,14 @@ import java.util.*
 
 data class Usuario(
     val id: Long,
-    val cpf: Int,
+    val cpf: Long,
     val nome: String,
     val nascimento: String,
     val genero: Boolean,
     val email: String,
     val endereco: String,
     val senha: String,
-    val telefone: Int
+    val telefone: Long
         ){
 
     var favoritos = 0
